Moved method test_streaming_rdt_nixl from 'test_streaming_generator.py' to 'test_gpu_object_nixl.py'
